Create DecoratedSuccessResponse
